added the enpoint performances (post,get,put,delete)

+++ endpoints/athletes.py
@@ -2,13 +2,14 @@
 #Importing the necessary libraries
 from fastapi import APIRouter,Depends, HTTPException
 from database import get_db,init_db
+from utils import get_current_user
 from pydantic import BaseModel
 import sqlite3
 from enum import Enum
 
 
 
-router=APIRouter(prefix="/athletes")
+app=APIRouter(prefix="/athletes")
 
 class GENDERENUM(str,Enum):
     male="male"
@@ -25,9 +26,12 @@ class AthleteSchema(BaseModel):
 
 
 #POST CREATE ATHLETE
-@router.post('/create')
-def create_athlete(athlete: AthleteSchema,db: sqlite3.Connection = Depends(get_db)):
+@app.post('/create')
+def create_athlete(athlete: AthleteSchema,db: sqlite3.Connection = Depends(get_db),current_user=Depends(get_current_user)):
     cursor=db.cursor()
+    role=current_user.role
+    if role !="coach":
+        raise HTTPException(status_code=401, detail="You are not allowed to perform this action")
     try:
         cursor.execute(
             "INSERT INTO athlete(name,gender,age,weight,height,user_id) VALUES(?,?,?,?,?,?)",
@@ -35,23 +39,28 @@ def create_athlete(athlete: AthleteSchema,db: sqlite3.Connection = Depends(get_d
         db.commit()
         # return {f"athlete no.{athlete_id:cursor.lastrowid} created sucessfully" }
         return {f"athlete name : {athlete.name} created sucessfully" }
-    except sqlite3.IntegrityError:
-        raise HTTPException(status_code=400, detail="User does not exist")
+    except sqlite3.IntegrityError as e:
+        raise HTTPException(status_code=400, detail="User does not exist") from e
     
 
 
 #Get athlete list 
-@router.get('/athletes')
-def get_athletes(db: sqlite3.Connection = Depends(get_db)):
+@app.get('/athletes')
+def get_athletes(db: sqlite3.Connection = Depends(get_db),current_user=Depends(get_current_user)):
+    role=current_user.role
+    if role !="coach":
+        raise HTTPException(status_code=401, detail="You are not allowed to perform this action")
     cursor = db.cursor()
     cursor.execute("SELECT * FROM athlete")
-    athletes = cursor.fetchall()
-    return athletes
+    return cursor.fetchall()
 
 
 #UPDATE ATHLETE
-@router.put('/update/<int:athlete_id>')
-def update_athlete(athlete_id: int, athlete: AthleteSchema, db: sqlite3.Connection = Depends(get_db)):
+@app.put('/update/<int:athlete_id>')
+def update_athlete(athlete_id: int, athlete: AthleteSchema, db: sqlite3.Connection = Depends(get_db), current_user=Depends(get_current_user)):
+    role=current_user.role
+    if role !="coach":
+        raise HTTPException(status_code=401, detail="You are not allowed to perform this action")
     cursor = db.cursor()
     cursor.execute("UPDATE athlete SET name=?, gender=?, age=?, weight=?, height=?, user_id=? WHERE athlete_id=?",
                    (athlete.name, athlete.gender, athlete.age, athlete.weight, athlete.height, athlete.user_id, athlete_id))
@@ -61,8 +70,11 @@ def update_athlete(athlete_id: int, athlete: AthleteSchema, db: sqlite3.Connecti
     return {f"Athlete no.{athlete_id} updated successfully"}
 
 #DELETE ATHLETE
-@router.delete('/delete/<int:athlete_id>')
-def delete_athlete(athlete_id: int, db: sqlite3.Connection = Depends(get_db)):
+@app.delete('/delete/<int:athlete_id>')
+def delete_athlete(athlete_id: int, db: sqlite3.Connection = Depends(get_db),current_user=Depends(get_current_user)):
+    role=current_user.role
+    if role !="coach":
+        raise HTTPException(status_code=401, detail="You are not allowed to perform this action")
     cursor = db.cursor()
     cursor.execute("DELETE FROM athlete WHERE athlete_id=?", (athlete_id,))
     db.commit()
